wmps3/wmps/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
pydantic==2.8.2
pymodbus==3.6.6
PyYAML==6.0.2
evdev==1.6.1
python-multipart==0.0.9
